# ballot-interpreter: Reuse computed Otsu level threshold