## Overview

Instead of recomputing the Otsu level when we score bubbles and write-ins, use the initial value computed after cropping the image. There are no further modifications to the image (other than rotation to normalize orientation), so this should be safe. This should save a little bit of time (according to the benchmarks, about 10%!)

## Demo Video or Screenshot
N/A

## Testing Plan
Updated existing benchmark saved data based on `main`, then ran the benchmarks again after the change to track improvement:

```
Blank HMPB interpretation
         runs   min       max       mean         median
------   ----   -------   -------   ----------   ------
old      50     69ms      118ms     74ms ± 3ms   72ms
new      50     61ms      71ms      65ms ± 1ms   65ms
change          -12.02%   -39.72%   -11.65%      -9.52%

Marked HMPB interpretation
         runs   min       max       mean         median
------   ----   -------   -------   ----------   -------
old      50     69ms      119ms     74ms ± 3ms   73ms
new      50     60ms      100ms     66ms ± 2ms   65ms
change          -13.69%   -15.48%   -10.76%      -10.61%
```
## Checklist

- [ ] I have added [logging](https://github.com/votingworks/vxsuite/tree/main/libs/logging) where appropriate to any new user actions, system updates such as file reads or storage writes, or errors introduced.
<!-- for user-facing changes, non-user facing changes can remove or ignore the below items -->
- [ ] I have added a screenshot and/or video to this PR to demo the change
- [ ] I have added the "user_facing_change" label to this PR to automate an announcement in #machine-product-updates